pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["marisiman <email_dari_marisiman@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
bcrypt = "4.1.2"
sqlalchemy = "^2.0.31"
pymysql = "1.1.0"
python-dotenv = "1.0.1"
flask-bcrypt = "1.0.1"
flask-cors = "4.0.0"
flask-jwt-extended = "4.6.0"
flask-sqlalchemy = "3.1.1"
poetry-dotenv-plugin = "^0.2.0"
psycopg2 = "^2.9.9"
flask-migrate = "^4.0.7"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
